Update 1.0.4.75
Update -
In this update you no longer have to copy the update zip to the SDA
directory and extract it, instead SDA will now download the update for
you along with extract it, (using Extract Update).
(SDA can now update itself if you allow)

Main -
Changed the login token entry to a label from a text field.

Theme -
Fixed a bug where in noTheme mode, leaving the theme editor would cause
a theme to be set
When you open the theme editor, it selects your currently saved theme.
In noTheme mode you can now enable the theme via the theme editor.
Changed the default theme to be a bit lighter.

:^)